+++ cctools/gprof/arcs.c
@@ -151,7 +151,7 @@ void)
 {
     nltype	*parentp, **timesortnlp;
     arctype	*arcp;
-    long	index;
+    unsigned long index;
 
 	/*
 	 * initialize various things:
@@ -244,7 +244,7 @@ void)
 	for(index = 0; index < nname; index++){
 	    timesortnlp[index] = &nl[index];
 	}
-	for(index = 1; index <= ncycle; index++){
+	for(index = 1; index <= (unsigned long)ncycle; index++){
 	    timesortnlp[nname + index - 1] = &cyclenl[index];
 	}
 	qsort(timesortnlp, nname + ncycle, sizeof(nltype *),
@@ -260,7 +260,7 @@ void
 dotime(
 void)
 {
-    int	index;
+    unsigned long index;
 
 	cycletime();
 	for(index = 0; index < nname; index += 1){
@@ -486,7 +486,7 @@ void
 doflags(
 void)
 {
-    int index;
+    long index;
     nltype *childp;
     nltype *oldhead;
