+++ cctools/otool/m88k_disasm.c
@@ -1132,7 +1132,7 @@ enum bool verbose)
 
 	if(reloc_found && r_extern == 1){
 	    if(symbols[r_symbolnum].n_un.n_strx < 0 ||
-	       symbols[r_symbolnum].n_un.n_strx >= strings_size)
+	       (unsigned long)symbols[r_symbolnum].n_un.n_strx >= strings_size)
 		name = "bad string offset";
 	    else
 		name = strings + symbols[r_symbolnum].n_un.n_strx;
